--- django_project/blog/custom_context_processor.py
@@ -1,14 +1,0 @@
-from .models import Category
-from django.db.models import Count
-
-
-def category_renderer(request):
-    category_qs = Category.objects.annotate(posts_count=Count("post"))
-    try:
-        current_category = request.resolver_match.kwargs["category"]
-    except (KeyError, AttributeError):
-        current_category = "None"
-    return {
-        "category_qs": category_qs,
-        "current_category": current_category,
-    }
