+++ django_project/blog/templates/blog/parts/base.html
@@ -48,6 +48,9 @@
         <div class="alert alert-{{ message.tags }}">{{ message }}</div>
         {% endfor %} {% endif %}
         <main id="main">
+          <div id="breadcrumbs-container">
+            {% include "blog/parts/breadcrumbs.html" %}
+          </div>
           {% block content %}
           {% endblock %}
         </main>
