+++ django_project/django_project/settings/base_settings.py
@@ -116,7 +116,8 @@ TEMPLATES = [
                 "django.template.context_processors.request",
                 "django.contrib.auth.context_processors.auth",
                 "django.contrib.messages.context_processors.messages",
-                "blog.custom_context_processor.category_renderer",
+                "blog.context_processors.category_renderer",
+                "blog.context_processors.breadcrumbs",
             ],
             "debug": True,
         },
